requirements.txt
# Core dependencies
discord.py>=2.0.0
requests>=2.28.0
python-dotenv>=1.0.0
colorama>=0.4.6
aiohttp>=3.8.3
asyncio>=3.4.3
Pillow>=9.4.0

# Database
peewee>=3.15.0

# Google API integration
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.78.0
google-auth-httplib2>=0.1.0

# OpenAI integration
openai>=1.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.20.0
aioresponses>=0.7.4


# For web viewing
flask>=2.0.0
